boilerplate

+++ app/src/main/java/com/metroair/job_card_management/data/local/database/entities/AssetEntity.kt
@@ -9,7 +9,7 @@ data class AssetEntity(
     val itemCode: String,
     val itemName: String,
     val category: String,
-    val currentStock: Int,
-    val minimumStock: Int,
+    val currentStock: Double,
+    val minimumStock: Double,
     val unitOfMeasure: String
 )
